move file to new folder and create __init__.py
