+++ stochatreat/__init__.py
@@ -1,0 +1,14 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Wed Jul 10 12:16:55 2019
+
+===============================================================================
+@author:    Manuel Martinez
+@project:   stochatreat
+===============================================================================
+"""
+
+__version__ = 1.0
+__author__ = 'Manuel Martinez'
+
+from .stochatreat import stochatreat
